Offline_Assignments/KG_Hybrid_RAG/requirements.txt
# Core UI
streamlit==1.41.1

# OpenAI
openai==1.60.2

# Knowledge graph
neo4j==5.26.0
graphiti==0.1.13
graphiti-core==0.7.8

# Environment management    
python-dotenv==1.0.1

# Hybrid retrieval
faiss-cpu==1.13.0
rank-bm25==0.2.2

# LangChain ecosystem
langchain-core==0.3.25
langchain-community==0.3.8
langchain-openai==0.2.10

# Image handling
pillow==10.4.0
